refactor: types and calculations
