+++ compact_test.go
